refactor: move components files into separate folders
