+++ src/components/Flyout/Flyout.tsx
@@ -1,9 +1,9 @@
 import React, { useContext } from 'react';
 import styles from './Flyout.module.css';
-import { ThemeContext } from '../pages/MainPage';
+import { ThemeContext } from '../../pages/MainPage';
 import { useDispatch, useSelector } from 'react-redux';
-import { RootState } from '../store/store';
-import { unselectAllItems } from '../store/reducers/itemsSlice';
+import { RootState } from '../../store/store';
+import { unselectAllItems } from '../../store/reducers/itemsSlice';
 
 interface FlyoutProps {
   downloadHandler: () => void;
