+++ src/components/ResultItem/ResultItem.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import { Provider as ReduxProvider } from 'react-redux';
 import ResultItem from './ResultItem';
-import { setupStore } from '../store/store';
+import { setupStore } from '../../store/store';
 
 const store = setupStore();
