+++ src/main.tsx
@@ -4,7 +4,7 @@ import App from './App.tsx';
 import './index.css';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import ErrorPage from './pages/ErrorPage.tsx';
-import ItemDetails from './components/ItemDetails.tsx';
+import ItemDetails from './components/ItemDetails/ItemDetails.tsx';
 
 const router = createBrowserRouter([
   {
